app/requirements.txt
numpy
Pillow
scapy